Aula 33 - Dialog de Confirmação

+++ crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -8,6 +8,7 @@ import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/err
 
 import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -71,21 +72,34 @@ export class CoursesComponent implements OnInit {
   }
 
   onDelete(course: Course) {
-      this.coursesService.delete(course._id).subscribe(
-        () => {
-          this.refresh();
-          this.snackBar.open('Curso deletado com sucesso!', 'X',
-          {
-            duration: 5000,
-            verticalPosition: 'top',
-            horizontalPosition: 'center'
-          });
-        },
-        error => {
-          this.onError("Erro ao tentar deletar curso!");
-        }
 
-      );
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover este curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.delete(course._id).subscribe(
+          () => {
+            this.refresh();
+            this.snackBar.open('Curso deletado com sucesso!', 'X',
+            {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center'
+            });
+          },
+          error => {
+            this.onError("Erro ao tentar deletar curso!");
+          }
+
+        );
+
+      }
+    });
+
+
+
   }
 
 }
